implementing manage listings functionality

+++ app/Http/Controllers/ListingController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Listing;
 use App\Models\User;
+use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class ListingController extends Controller
@@ -49,7 +50,7 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        $formFields['user_id'] = auth()->id();
+        $formFields['user_id'] = Auth::id();
 
         Listing::create($formFields);
 
@@ -65,6 +66,11 @@ class ListingController extends Controller
     // Update Listing Data
     public function update(Request $request, Listing $listing)
     {
+        //make sure that the logged in user is the owner
+        if ($listing->user_id != Auth::user()->id) {
+            abort(403, "Unauthorized action");
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -87,10 +93,29 @@ class ListingController extends Controller
     // Delete Listing
     public function destroy(Listing $listing)
     {
+        //make sure that the logged in user is the owner
+        if ($listing->id != Auth::user()->id) {
+            abort(403, "Unauthorized action");
+        }
+
         if ($listing->logo && Storage::disk('public')->exists($listing->logo)) {
             Storage::disk('public')->delete($listing->logo);
         }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        $listings = [];
+
+        if (Auth::check()) {
+            $user = Auth::user();
+
+            $listings = $user->is_admin ? Listing::all() : $user->listings;
+        }
+
+        return view('listings.manage', ['listings' => $listings]);
     }
 }
